+++ src/main/java/com/Merlin/Inventory/Management/System/Transaction/TransactionItemRequest.java
@@ -2,6 +2,6 @@ package com.Merlin.Inventory.Management.System.Transaction;
 
 public record TransactionItemRequest(
         Long productId,
-        int quantity
+        Double quantity
 ) {
 }
